Added helper functions for now and then the screen files
